Add support for authentication and fractional CPU allocation
Includes:
- Integration of API key-based authentication using Ktor's authentication feature.
- Migration from `cpuCount` (integer) to `cpus` (double) to support fractional CPU allocation.
- Added `DockerClientProvider` for centralized Docker client configuration.
- Improved Docker container security settings (e.g., sandboxing with `readonlyRootfs` and `no-new-privileges`).
- Updated REST API, database schema, and data models to reflect CPU changes.
- Configuration updates in `application.yaml` for API key and Docker host.

+++ src/main/kotlin/com/example/executor/ExecutionResponse.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 data class ExecutionResponse(
     val id: String,
     val script: String,
-    val cpuCount: Int,
+    val cpuCount: Double,
     val memoryMb: Int,
     val status: ExecutionStatus,
     val executorId: String? = null,
